refactor: move all action creators for room types to rooms action creators file
